% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDesc.R
\docType{data}
\name{cardSim}
\alias{cardSim}
\title{Simulated cardiac medication data for survival analysis}
\format{
A data frame with 1000 rows and 6 columns:
\describe{
 \item{id}{Unique identifier for each row.}
 \item{time}{Time in years to the outcome specified by status.}
 \item{status}{1 if the patient experienced cardiac arrest. 0 if censored before that.}
 \item{drug}{Medication the patient received for cardiac health (drug 1 or drug 0).}
 \item{age}{Age of the patient, ranges from 18 to 65 years.}
 \item{bmi}{Patient body mass index. Majority of observations fall between 22 and 30.}
}
}
\description{
This dataset was created to demonstrate the effects of Local Control on correcting bias within a set of data.
}
\author{
Lauve NR, Lambert CG
}
\keyword{data}
