% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDesc.R
\docType{data}
\name{framingham}
\alias{framingham}
\title{Framingham heart study data extract on smoking and hypertension.}
\format{
A data frame with 2316 rows and 11 columns:
\describe{
 \item{female}{Sex of the patient. 1=female, 0=male.}
 \item{totchol}{Total cholesterol of patient at study entry.}
 \item{age}{Age of the patient at study entry.}
 \item{bmi}{Patient body mass index.}
 \item{BPVar}{Average units of systolic and diastolic blood pressure above normal: ((SystolicBP-120)/2) + (DiasystolicBP-80)}
 \item{heartrte}{Patient heartrate taken at study entry.}
 \item{glucose}{Patient blood glucose level.}
 \item{cursmoke}{Whether or not the patient was a smoker at the time of study entry.}
 \item{outcome}{Did the patient die, experience hypertension, or leave the study without experiencing either event.}
 \item{time_outcome}{The time at which the patient experienced outcome.}
 \item{cigpday}{Number of cigarettes smoked per day at time of study entry.}
}
}
\description{
Data collected over a 24 year study suitable for competing
  risks survival analysis of hypertension and death as a function of smoking.
}
\references{
\itemize{
    \item Dawber TR, Meadors GF, Moore FE Jr. Epidemiological approaches to heart disease: the Framingham Study. Am J Public Health Nations Health. 1951;41(3):279-281.
    \item Teaching Datasets - Public Use Datasets. \url{https://biolincc.nhlbi.nih.gov/teaching/}.
}
}
\keyword{data}
