% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDesc.R
\docType{data}
\name{lindner}
\alias{lindner}
\title{Lindner Center for Research and Education study on Abciximab cost-effectiveness and survival}
\format{
A data frame with 996 rows and 10 columns:
\describe{
 \item{lifepres}{Life years preserved post treatment: 0 (died) vs. 11.6 (survived).}
 \item{cardbill}{Cardiac related billing in dollars within 12 months.}
 \item{abcix}{Indicates whether the patient received Abciximab treatment: 1=yes 0=no.}
 \item{stent}{Was a stent depolyed? 1=yes, 0=no.}
 \item{height}{Patient height in centimeters.}
 \item{female}{Patient sex: 1=female, 0=male.}
 \item{diabetic}{Was the patient diabetic? 1=yes, 0=no.}
 \item{acutemi}{Had the patient suffered an acute myocardial infarction witih the last seven days? 1=yes, 0=no.}
 \item{ejecfrac}{Left ventricular ejection fraction.}
 \item{ves1proc}{Number of vessels involved in the first PCI procedure.}
}
}
\description{
The effects of Abciximab use on both survival and cardiac billing.
}
\references{
Kereiakes DJ, Obenchain RL, Barber BL, Smith A, McDonald M, Broderick TM, Runyon JP, Shimshak TM, Schneider JF, Hattemer CR, Roth EM, Whang DD, Cocks D, Abbottsmith CW. Abciximab provides cost-effective survival advantage in high-volume interventional practice. Am Heart J. 2000;140(4):603-610.
}
\keyword{data}
