% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedLC.R
\docType{package}
\name{LocalControl-deprecated}
\alias{LocalControl-deprecated}
\alias{localControlNearestNeighbors}
\alias{localControlCompetingRisks}
\alias{plotLocalControlCIF}
\alias{plotLocalControlLTD}
\title{Deprecated LocalControl functions}
\description{
These functions are provided for compatibility with previous versions of LocalControl.
They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{localControlNearestNeighbors} \tab Now called using \code{\link{LocalControl}} with the outcomeType = "cross-sectional". \cr
  \code{localControlCompetingRisks}   \tab Now called using \code{\link{LocalControl}} with the outcomeType = "survival". \cr
  \code{plotLocalControlCIF}   \tab Now called using \code{\link{plot.LocalControlCR}}. \cr
  \code{plotLocalControlLTD}   \tab Now called using \code{\link{plot.LocalControlCS}}. \cr
}
}

