% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localControl.R
\name{LocalControlCompetingRisksConfidence}
\alias{LocalControlCompetingRisksConfidence}
\title{Calculate confidence intervals around the cumulative incidence functions (CIFs) generated by LocalControl when outcomeType = "survival".}
\usage{
LocalControlCompetingRisksConfidence(
  LCCompRisk,
  confLevel = "95\%",
  confTransform = "asin"
)
}
\arguments{
\item{LCCompRisk}{Output from a successful call to LocalControl with outcomeType = "survival".}

\item{confLevel}{Level of confidence with which the confidence intervals will be formed. Choices are: "90\%", "95\%", "98\%", "99\%".}

\item{confTransform}{Transformation of the confidence intervals, defaults to arcsin ("asin").
"log" and "linear" are also implemented.}
}
\description{
Given the output of \code{\link{LocalControl}}, this function produces pointwise standard error estimates
for the cumulative incidence functions (CIFs) using a modified version of Choudhury's approach (2002). This function currently supports
the creation of 90\%, 95\%, 98\%, and 99\% confidence intervals with linear, log(-log), and arcsine transformations of the estimates.
}
\examples{
 data(cardSim)
 results = LocalControl(data = cardSim,
                        outcomeType = "survival",
                        outcomeColName = "status",
                        timeColName = "time",
                        treatmentColName = "drug",
                        treatmentCode = 1,
                        clusterVars = c("age", "bmi"))

 conf = LocalControlCompetingRisksConfidence(results)

}
\references{
\itemize{
   \item Lauve NR, Nelson SJ, Young SS, Obenchain RL, Lambert CG. LocalControl: An R Package for Comparative Safety and Effectiveness Research. Journal of Statistical Software. 2020. p. 1–32. Available from: http://dx.doi.org/10.18637/jss.v096.i04
   \item Choudhury JB (2002) Non-parametric confidence interval estimation for competing risks analysis: application to contraceptive data. Stat Med 21:1129-1144. doi: 10.1002/sim.1070
}
}
